Move base html to ui app
