+++ loaders/templates/loaders/loaderfailure_detail.html
@@ -1,4 +1,4 @@
-{% extends "radio/base.html" %}
+{% extends "ui/base.html" %}
 
 {% load radio %}
