+++ loaders/templates/loaders/loaderfailure_list.html
@@ -1,4 +1,4 @@
-{% extends "radio/base.html" %}
+{% extends "ui/base.html" %}
 
 {% block title %}Failures{% endblock %}
