+++ radio/templates/radio/radio_stats.html
@@ -1,4 +1,4 @@
-{% extends "radio/base.html" %}
+{% extends "ui/base.html" %}
 
 {% load static %}
