+++ ui/templates/ui/index.html
@@ -1,4 +1,4 @@
-{% extends "radio/base.html" %}
+{% extends "ui/base.html" %}
 
 {% block title %}Home{% endblock %}
